oca_bot/requirements.txt
flask==2.2.3
openai==0.27.2
python-dotenv==1.0.0
line-bot-sdk==2.4.2
